+++ src/stories/BackButton.stories.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
+import { withKnobs, text, object } from '@storybook/addon-knobs';
 
 import BackButton from '../components/BackButton';
 
@@ -10,6 +11,7 @@ const actions = {
 export default {
     title: 'Components/BackButton',
     decorators: [
+        withKnobs,
         story => <div style={{ position: 'absolute', left: '5%', top: '10%' }}>{story()}</div>
     ],
     component: BackButton